refactor and fix host update
