+++ services/host.go
@@ -3,8 +3,9 @@ package services
 import (
 	"context"
 	"log"
+	"strings"
 
-	"github.com/PUMATeam/catapult/node"
+	"github.com/PUMATeam/catapult/util"
 
 	"github.com/PUMATeam/catapult/model"
 	"github.com/PUMATeam/catapult/repositories"
@@ -21,21 +22,20 @@ type hostsService struct {
 	hostsRepository repositories.Hosts
 }
 
-func (as *hostsService) HostByID(ctx context.Context, id uuid.UUID) (*model.Host, error) {
-	return as.hostsRepository.HostByID(ctx, id)
+func (hs *hostsService) HostByID(ctx context.Context, id uuid.UUID) (*model.Host, error) {
+	return hs.hostsRepository.HostByID(ctx, id)
 }
 
-func (as *hostsService) ListHosts(ctx context.Context) ([]model.Host, error) {
-	return as.hostsRepository.ListHosts(ctx)
+func (hs *hostsService) ListHosts(ctx context.Context) ([]model.Host, error) {
+	return hs.hostsRepository.ListHosts(ctx)
 }
 
-// TODO: not working
-func (as *hostsService) UpdateHostStatus(ctx context.Context, host *model.Host, status int) error {
+func (hs *hostsService) UpdateHostStatus(ctx context.Context, host model.Host, status int) error {
 	host.Status = status
-	return as.hostsRepository.UpdateHost(ctx, host)
+	return hs.hostsRepository.UpdateHost(ctx, host)
 }
 
-func (as *hostsService) AddHost(ctx context.Context, newHost NewHost) (uuid.UUID, error) {
+func (hs *hostsService) AddHost(ctx context.Context, newHost NewHost) (uuid.UUID, error) {
 	// TODO add validations and rollback in case ansible fails
 	// also, run within a worker pool framework
 	host := model.Host{
@@ -47,18 +47,42 @@ func (as *hostsService) AddHost(ctx context.Context, newHost NewHost) (uuid.UUID
 		Password: newHost.Password,
 	}
 
-	id, err := as.hostsRepository.AddHost(ctx, host)
-	n := node.NewNodeService(host)
-	err = n.InstallHost()
+	id, err := hs.hostsRepository.AddHost(ctx, host)
+	host.ID = id
+	err = hs.InstallHost(host)
 	if err != nil {
-		log.Println("Updating status of host", host.Name, "to up")
-		err = as.UpdateHostStatus(ctx, &host, UP)
-		if err != nil {
-			log.Println("Failed to update status of host", host.Name, "to up")
-		}
+		return uuid.Nil, err
+	}
+
+	log.Println("Updating status of host", host.Name, "to up")
+	err = hs.UpdateHostStatus(ctx, host, UP)
+	if err != nil {
+		log.Println("Failed to update status of host", host.Name, "to up")
 	}
 
 	return id, err
+}
+
+// InstallHost installs prerequisits on the host
+// TODO: leaving it as public to allow a user add a host
+// without installing right away
+func (hs *hostsService) InstallHost(h model.Host) error {
+	hi := hostInstall{
+		User:            h.User,
+		FcVersion:       FcVersion,
+		AnsiblePassword: h.Password,
+	}
+	ac := util.NewAnsibleCommand(util.SetupHostPlaybook,
+		h.User,
+		h.Address,
+		util.StructToMap(hi, strings.ToLower))
+	err := ac.ExecuteAnsible()
+	if err != nil {
+		log.Println("Error during host install: ", err)
+		return err
+	}
+
+	return nil
 }
 
 type NewHost struct {
@@ -74,3 +98,12 @@ const (
 	INSTALLING int = 2
 	UP         int = 3
 )
+
+type hostInstall struct {
+	User            string `json:"ignore"`
+	AnsiblePassword string `json:"ansible_ssh_pass"`
+	FcVersion       string
+}
+
+// TODO make it configurable
+const FcVersion = "0.15.0"
